Fix variable "killer" as global variable in The Entrance the level
When an enemy outside of the pyramid is killed, the local variable "killer" initialized to true.
But when the sector loads, such as teleporting or spawning, the non-existent variable "killer" is undefined behavior which will read as false.
This fix places a one-time script trigger at the level start to initialize the cross-sector global variable "state.killer".